# ⚡️ Simplify and speed up `SEARCH` response parsing