The `mailbox-data` `SEARCH` response parser was refactored, as part of the work on `CONDSTORE` support.  This commit contains only the refactoring, which (coincidentally) brings a significant speed improvement:

```
invalid_search_response_multiple_result_with_trailing_space
                v0.4.6-2-g3ac9912:     70865.1 i/s
                            0.4.6:     49505.1 i/s - 1.43x  slower

invalid_search_response_single_result_with_trailing_space
                v0.4.6-2-g3ac9912:     74398.0 i/s
                            0.4.6:     67791.1 i/s - 1.10x  slower

rfc3501_7.2.5_SEARCH_response_example
                v0.4.6-2-g3ac9912:     65968.9 i/s
                            0.4.6:     53490.9 i/s - 1.23x  slower

search_response_multiple_seq_nums_returned
                v0.4.6-2-g3ac9912:     62777.4 i/s
                            0.4.6:     47852.3 i/s - 1.31x  slower

search_response_single_seq_nums_returned
                v0.4.6-2-g3ac9912:     79656.5 i/s
                            0.4.6:     69234.9 i/s - 1.15x  slower

search_response_with_condstore_modseq
                v0.4.6-2-g3ac9912:     55376.4 i/s
                            0.4.6:     39129.0 i/s - 1.42x  slower
```